Quadcopter.update() method completed. Physics divide by (fps=0) fixed.
Added function to prevent physics.fps from being 
less than a threshold, to prevent a divide by zero 
when calculating accelerations and velocities.
There was a bug previously because in the first 
frame of the program, p5js returns an fps of 0.
Added a physics.timeScale factor to speed or 
slow the physics of the sim.

Note that the apparent speed is a combination of 
physics.pixelsPerMeter and
physics.framesPerSimSecond, which is now scaled 
by physics.timeScale.